review comments; fix tests
